# api/openapi.yaml
openapi: 3.0.2
info:
  description: |
    Handles all identities for tracking the users of the Moov platform.
  title: Moov Identity API
  version: 0.0.1
servers:
- description: Local Testing
  url: https://local.moov.io/
- description: Production
  url: https://api.moov.io/
- description: Internal address of service
  url: https://identity.srvs.moov.local/
paths:
  /invites:
    get:
      parameters:
      - description: Filter in only for specific Organization
        explode: true
        in: query
        name: orgID
        required: false
        schema:
          $ref: '#/components/schemas/UUID'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Invite'
                maxItems: 300
                type: array
          description: Invites that are outstanding
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: List outstanding invites
      tags:
      - invites
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invite'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invite'
          description: Invite sent
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: Send an email invite to a new user
      tags:
      - invites
  /invites/{inviteID}:
    delete:
      parameters:
      - description: ID of the invite to delete
        explode: false
        in: path
        name: inviteID
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "204":
          description: Invite was removed
        "404":
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: Delete an invite that was sent and invalidate the token.
      tags:
      - invites
  /identities:
    get:
      parameters:
      - description: Filter in only for specific Organization
        explode: true
        in: query
        name: orgID
        required: false
        schema:
          $ref: '#/components/schemas/UUID'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Identity'
                maxItems: 300
                type: array
          description: List of identities/users in the system
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: List identities and associates userId
      tags:
      - identities
  /identities/{identityID}:
    get:
      parameters:
      - description: ID of the Identity to lookup
        explode: false
        in: path
        name: identityID
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
          description: List of identities/users in the system
        "404":
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: List identities and associates userId
      tags:
      - identities
    put:
      parameters:
      - description: ID of the Identity to lookup
        explode: false
        in: path
        name: identityID
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Identity'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Identity'
          description: Identity was updated.
        "404":
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: Update a specific Identity
      tags:
      - identities
  /identities/{identityID}/credentials:
    get:
      parameters:
      - description: ID of the Identity to lookup
        explode: false
        in: path
        name: identityID
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Credential'
                maxItems: 300
                type: array
          description: List of credentials tied to this identity
        "404":
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - GatewayAuth: []
      summary: List the credentials this user has used.
      tags:
      - identities
  /login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        description: Arguments needed to match up the OIDC user data with a user in
          the system
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggedIn'
          description: User successfully logged in.
        "404":
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - ServiceAuth: []
      summary: "Complete a login via a OIDC.\nOnce the OIDC client service has authenticated\
        \ their identity the client service will call \nthis endpoint to record and\
        \ finish the login to get their token to use the API.\n\nIf the client service\
        \ recieves a 404 they must send them to registration if its allowed per the\
        \ client or check for an invite for authenticated users email before sending\
        \ to registration.      \n"
      tags:
      - internal
  /register:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
        description: Arguments needed register a user with OIDC credentials.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggedIn'
          description: User successfully registered, they can now log in
        default:
          content:
            text/plain:
              schema:
                maxLength: 0
                pattern: //i
                type: string
          description: Empty response for unauthorized or any other returned http
            status code
      security:
      - ServiceAuth: []
      summary: "Register user based on OIDC credentials. \nThis is called by the OIDC\
        \ client services we create to register the user with what \navailable information\
        \ they have and obtain from the user.\n"
      tags:
      - internal
components:
  responses:
    Empty:
      content:
        text/plain:
          schema:
            maxLength: 0
            pattern: //i
            type: string
      description: Empty response for unauthorized or any other returned http status
        code
  schemas:
    UUID:
      description: UUID v4
      format: uuid
      maxLength: 36
      pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
      type: string
    Email:
      description: Email Address
      example: john.doe@gmail.com
      format: email
      maxLength: 255
      pattern: ^([a-zA-Z0-9_\-\.]+)@([a-zA-Z0-9_\-\.]+)\.([a-zA-Z]{2,5})$
      type: string
    DateTime:
      format: date-time
      maxLength: 24
      type: string
    SubjectID:
      description: ID of the remote OIDC server gives to this identity
      example: aSDDfasd23Dfasdf
      maxLength: 255
      pattern: ^[0-9a-zA-Z_-]+$
      readOnly: true
      type: string
    Provider:
      description: OIDC provider that was used to handle authentication of this user.
      example: google
      maxLength: 255
      pattern: ^\w+$
      readOnly: true
      type: string
    Invite:
      additionalProperties: false
      description: Describes an invite that was sent to a user to join.
      example:
        invitedBy: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        redeemed: true
        inviteID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        tenantID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        expiresOn: 2000-01-23T04:56:07.000+00:00
        invitedOn: 2000-01-23T04:56:07.000+00:00
        email: john.doe@gmail.com
      properties:
        inviteID:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        tenantID:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        email:
          description: Email Address
          example: john.doe@gmail.com
          format: email
          maxLength: 255
          pattern: ^([a-zA-Z0-9_\-\.]+)@([a-zA-Z0-9_\-\.]+)\.([a-zA-Z]{2,5})$
          type: string
        invitedBy:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        invitedOn:
          format: date-time
          maxLength: 24
          type: string
        expiresOn:
          format: date-time
          maxLength: 24
          type: string
        redeemed:
          description: If this invite has been redeemed for a registration
          readOnly: true
          type: boolean
      type: object
    Identity:
      additionalProperties: false
      description: |
        Properties of an Identity. These users will under-go KYC checks thus all the information
      example:
        lastName: Doe
        lastLogin:
          credentialId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          "on": 2000-01-23T04:56:07.000+00:00
        addresses:
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        nickName: nickName
        phones:
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        suffix: suffix
        birthDate: 2020-12-31T00:00:00.000+0000
        firstName: John
        emailVerified: true
        identityID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        middleName: Jimmy
        registeredOn: 2000-01-23T04:56:07.000+00:00
        email: john.doe@gmail.com
        status: rejected
      properties:
        identityID:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        firstName:
          example: John
          maxLength: 255
          pattern: ^\w+$
          type: string
        middleName:
          example: Jimmy
          maxLength: 255
          pattern: ^\w+$
          type: string
        lastName:
          example: Doe
          maxLength: 255
          pattern: ^\w+$
          type: string
        nickName:
          maxLength: 255
          nullable: true
          pattern: ^\w+$
          type: string
        suffix:
          maxLength: 255
          nullable: true
          pattern: ^\w+$
          type: string
        birthDate:
          example: 2020-12-31
          format: date
          maxLength: 10
          type: string
        status:
          enum:
          - rejected
          - none
          - review-required
          - kyc
          - ofac
          - cip
          type: string
        email:
          description: Email Address
          example: john.doe@gmail.com
          format: email
          maxLength: 255
          pattern: ^([a-zA-Z0-9_\-\.]+)@([a-zA-Z0-9_\-\.]+)\.([a-zA-Z]{2,5})$
          type: string
        emailVerified:
          description: The user has verified they have access to this email
          readOnly: true
          type: boolean
        phones:
          items:
            $ref: '#/components/schemas/Phone'
          maxItems: 300
          type: array
        addresses:
          items:
            $ref: '#/components/schemas/Address'
          maxItems: 300
          type: array
        registeredOn:
          format: date-time
          maxLength: 24
          type: string
        lastLogin:
          $ref: '#/components/schemas/LastLogin'
      required:
      - email
      - firstName
      - lastName
      type: object
    LastLogin:
      additionalProperties: false
      description: |
        Defines when and what credential was used for the last login
      example:
        credentialId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        "on": 2000-01-23T04:56:07.000+00:00
      properties:
        credentialId:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        "on":
          format: date-time
          maxLength: 24
          type: string
      type: object
    Credential:
      additionalProperties: false
      description: |
        Description of a successful OpenID connect credential
      example:
        lastUsedOn: 2000-01-23T04:56:07.000+00:00
        provider: google
        identityID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        credentialID: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        createdOn: 2000-01-23T04:56:07.000+00:00
        subjectID: aSDDfasd23Dfasdf
        enabled: true
      properties:
        credentialID:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        provider:
          description: OIDC provider that was used to handle authentication of this
            user.
          example: google
          maxLength: 255
          pattern: ^\w+$
          readOnly: true
          type: string
        subjectID:
          description: ID of the remote OIDC server gives to this identity
          example: aSDDfasd23Dfasdf
          maxLength: 255
          pattern: ^[0-9a-zA-Z_-]+$
          readOnly: true
          type: string
        identityID:
          description: UUID v4
          format: uuid
          maxLength: 36
          pattern: ^[0-9a-fA-F]{8}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{4}\-[0-9a-fA-F]{12}$
          type: string
        enabled:
          description: If disabled user will be unable to use this method of authentication
            without help from support removing the lock.
          type: boolean
        createdOn:
          format: date-time
          maxLength: 24
          type: string
        lastUsedOn:
          format: date-time
          maxLength: 24
          type: string
      type: object
    OFACSearch:
      properties:
        entityId:
          description: SDN EntityID of the Entity
          example: "2141"
          type: string
        sdnName:
          description: Name of the SDN entity
          example: John Smith
          type: string
        sdnType:
          description: SDN entity type
          example: Individual
          type: string
        match:
          description: Percentage of similarity between the Customer name and this
            OFAC entity
          example: 0.91
          type: number
        createdAt:
          example: 2016-08-29T09:12:33.001Z
          format: date-time
          type: string
      type: object
    Phone:
      additionalProperties: false
      description: Phone number
      example:
        valid: true
        number: number
        type: home
      properties:
        number:
          format: phone
          maxLength: 15
          pattern: ^[0-9]+$
          type: string
        valid:
          type: boolean
        type:
          enum:
          - home
          - mobile
          - work
          type: string
      type: object
    Address:
      additionalProperties: false
      description: Address of the Identity
      example:
        country: US
        validated: true
        address2: address2
        city: Cedar Falls
        address1: 111 Main St.
        postalCode: "50613"
        active: true
        state: IA
        type: primary
      properties:
        type:
          enum:
          - primary
          - secondary
          type: string
        address1:
          example: 111 Main St.
          maxLength: 255
          pattern: ^\w+$
          type: string
        address2:
          maxLength: 255
          nullable: true
          pattern: ^\w*$
          type: string
        city:
          example: Cedar Falls
          maxLength: 255
          pattern: ^\w+$
          type: string
        state:
          example: IA
          maxLength: 2
          minLength: 2
          pattern: ^\w+$
          type: string
        postalCode:
          example: "50613"
          maxLength: 5
          minLength: 5
          pattern: ^\d+$
          type: string
        country:
          example: US
          maxLength: 2
          minLength: 2
          pattern: ^\w+$
          type: string
        validated:
          type: boolean
        active:
          type: boolean
      type: object
    Login:
      additionalProperties: false
      description: |
        User has authenticated with an OIDC now to verify the system can find them. This verifies theirs stored credentials for a user that match the arguments
      example:
        provider: google
        subjectID: aSDDfasd23Dfasdf
      properties:
        provider:
          description: OIDC provider that was used to handle authentication of this
            user.
          example: google
          maxLength: 255
          pattern: ^\w+$
          readOnly: true
          type: string
        subjectID:
          description: ID of the remote OIDC server gives to this identity
          example: aSDDfasd23Dfasdf
          maxLength: 255
          pattern: ^[0-9a-zA-Z_-]+$
          readOnly: true
          type: string
      type: object
    LoggedIn:
      additionalProperties: false
      description: User has logged in and is being given a token to proof identity
      example:
        jwt: jwt
      properties:
        jwt:
          description: JWT token that provides authentication of identity
          maxLength: 4000
          pattern: ^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+/=]*$
          type: string
      type: object
    Register:
      additionalProperties: false
      description: Request to register a user in the system
      example:
        firstName: John
        lastName: Doe
        addresses:
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        - country: US
          validated: true
          address2: address2
          city: Cedar Falls
          address1: 111 Main St.
          postalCode: "50613"
          active: true
          state: IA
          type: primary
        provider: google
        nickName: nickName
        phones:
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        - valid: true
          number: number
          type: home
        middleName: Jimmy
        suffix: suffix
        birthDate: 2020-12-31T00:00:00.000+0000
        subjectID: aSDDfasd23Dfasdf
        email: john.doe@gmail.com
      properties:
        provider:
          description: OIDC provider that was used to handle authentication of this
            user.
          example: google
          maxLength: 255
          pattern: ^\w+$
          readOnly: true
          type: string
        subjectID:
          description: ID of the remote OIDC server gives to this identity
          example: aSDDfasd23Dfasdf
          maxLength: 255
          pattern: ^[0-9a-zA-Z_-]+$
          readOnly: true
          type: string
        firstName:
          example: John
          maxLength: 255
          pattern: ^\w+$
          type: string
        middleName:
          example: Jimmy
          maxLength: 255
          pattern: ^\w+$
          type: string
        lastName:
          example: Doe
          maxLength: 255
          pattern: ^\w+$
          type: string
        nickName:
          maxLength: 255
          nullable: true
          pattern: ^\w+$
          type: string
        suffix:
          maxLength: 255
          nullable: true
          pattern: ^\w+$
          type: string
        birthDate:
          example: 2020-12-31
          format: date
          maxLength: 10
          type: string
        email:
          description: Email Address
          example: john.doe@gmail.com
          format: email
          maxLength: 255
          pattern: ^([a-zA-Z0-9_\-\.]+)@([a-zA-Z0-9_\-\.]+)\.([a-zA-Z]{2,5})$
          type: string
        phones:
          items:
            $ref: '#/components/schemas/Phone'
          maxItems: 50
          type: array
        addresses:
          items:
            $ref: '#/components/schemas/Address'
          maxItems: 50
          type: array
      type: object
  securitySchemes:
    GatewayAuth:
      bearerFormat: JWT
      description: JWT that comes from the gateway that validates against the gateways
        public RSA key
      scheme: bearer
      type: http
    ServiceAuth:
      description: Hash of HMAC(X-Timestamp +'|' + X-Nonce, Secret)
      in: header
      name: X-Request-Hash
      type: apiKey