# pkg/server/api_invites.go
/*
 * Moov Identity API
 *
 * Handles all identities for tracking the users of the Moov platform. 
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package identityserver

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// A InvitesApiController binds http requests to an api service and writes the service results to the http response
type InvitesApiController struct {
	service InvitesApiServicer
}

// NewInvitesApiController creates a default api controller
func NewInvitesApiController(s InvitesApiServicer) Router {
	return &InvitesApiController{ service: s }
}

// Routes returns all of the api route for the InvitesApiController
func (c *InvitesApiController) Routes() Routes {
	return Routes{ 
		{
			"InvitesGet",
			strings.ToUpper("Get"),
			"/invites",
			c.InvitesGet,
		},
		{
			"InvitesInviteIDDelete",
			strings.ToUpper("Delete"),
			"/invites/{inviteID}",
			c.InvitesInviteIDDelete,
		},
		{
			"InvitesPost",
			strings.ToUpper("Post"),
			"/invites",
			c.InvitesPost,
		},
	}
}

// InvitesGet - List outstanding invites
func (c *InvitesApiController) InvitesGet(w http.ResponseWriter, r *http.Request) { 
	query := r.URL.Query()
	orgID := query.Get("orgID")
	result, err := c.service.InvitesGet(orgID)
	if err != nil {
		w.WriteHeader(500)
		return
	}
	
	EncodeJSONResponse(result, nil, w)
}

// InvitesInviteIDDelete - Delete an invite that was sent and invalidate the token.
func (c *InvitesApiController) InvitesInviteIDDelete(w http.ResponseWriter, r *http.Request) { 
	params := mux.Vars(r)
	inviteID := params["inviteID"]
	result, err := c.service.InvitesInviteIDDelete(inviteID)
	if err != nil {
		w.WriteHeader(500)
		return
	}
	
	EncodeJSONResponse(result, nil, w)
}

// InvitesPost - Send an email invite to a new user
func (c *InvitesApiController) InvitesPost(w http.ResponseWriter, r *http.Request) { 
	invite := &Invite{}
	if err := json.NewDecoder(r.Body).Decode(&invite); err != nil {
		w.WriteHeader(500)
		return
	}
	
	result, err := c.service.InvitesPost(*invite)
	if err != nil {
		w.WriteHeader(500)
		return
	}
	
	EncodeJSONResponse(result, nil, w)
}