# pkg/server/model_address.go
/*
 * Moov Identity API
 *
 * Handles all identities for tracking the users of the Moov platform. 
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package identityserver

// Address - Address of the Identity
type Address struct {

	Type string `json:"type,omitempty"`

	Address1 string `json:"address1,omitempty"`

	Address2 *string `json:"address2,omitempty"`

	City string `json:"city,omitempty"`

	State string `json:"state,omitempty"`

	PostalCode string `json:"postalCode,omitempty"`

	Country string `json:"country,omitempty"`

	Validated bool `json:"validated,omitempty"`

	Active bool `json:"active,omitempty"`
}