# pkg/server/model_phone.go
/*
 * Moov Identity API
 *
 * Handles all identities for tracking the users of the Moov platform. 
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package identityserver

// Phone - Phone number
type Phone struct {

	Number string `json:"number,omitempty"`

	Valid bool `json:"valid,omitempty"`

	Type string `json:"type,omitempty"`
}